Update to only take top 100 credits for rule 3
Rule 3 looks at the preponderance of credits in the various classification categories. The
student has to achieve more than 50% of the relevant credits (i.e. the credits used in the
calculations) - after discounting the worst 20 credits at Level 5 and level 6 - from level 5 and
6 combined in a classification band or higher to be awarded that classification. To carry out
the calculation, the Board must start with the highest classification and keep applying the rule
until the student meets the condition. Student A has 200 relevant credits so must achieve
more than 100 in the classification category or better for the Board to recommend that
particular classification.

p.s I do not know what I am doing with the code so it may need more updating I just noticed that you missed only taking the top 50% relevant credits (100) when testing it